Attach task status to the batch exit status so we can recover it for monitoring
Don't know why this didn't occur to me earlier. I've spent so much time fighting the listener behavior in batch trying to force it to raise an event after a stage even if one of the tasks failed. It's simpler just to not worry about that end stage event (in the case of a failure) but report enough detail from the failed task.