Added the concept of a 'read account override' to the front50 configuration.
ie:
- name: front50
   group: metadata
   baseUrl: http://front50.prod.netflix.net
   config:
     readAccountOverride: global

It's used by the ApplicationService to specify that a single account should be queried for application metadata, rather than the default behavior of aggregating across all available accounts.

A discussion is perhaps warranted on how best to wire in the 'global' registry long-term (right now we specifically only want reads to be overridden).

Fixes: https://github.com/spinnaker/gate/issues/13