Restructured config and fixed integration test
This was an absolute horror to fix. Trying to inject Spock mocks into Spring integration tests is a nightmare because the lifecycles don't match – Spock mocks can only be created in the scope of a feature method's execution. If any beans dependent on the things you want to mock are present in configuration classes they will already have been loaded causing NoSuchBeanDefinition all up in your grill. In the end I've used Groovy map coercion for the things I wanted to stub out. Not ideal as there's no flexibility in the data returned.