replace template params before json executes
Right now, many aspects of Orca orchestrations are precomputed before a pipeline is executed, this change will preprocess a pipeline based on trigger ( and upcoming pipeline parameters ).

This in effect will allow you to insert a regions array, for example, and actually have it work.

The immediate need for this is to fix highlander strategies not working with developer branch pipelines