Adds scale_to_cluster as an option for ResizeServerGroup.
The targeted ServerGroup is sized to the largest value for min, max, and desired capacity among
all ServerGroups in the cluster in the targeted ServerGroup's location.
The computed size can be increasted by either scaleNum or scalePct options up to but not exceeding
the computed new capacity 'max'.
Adds scale_exact as an explicit option for scaling to a configured capacity.
adds ResizeStrategy implementations for various handlers of resizeActions
converts CloudProviderAwareSupport to CloudProviderAware interface with default methods
 side-effect of needing to remove a couple @CompileStatic annotations due to our poor choice of generic for Stage
adds targetLocation as a context attribute when generating static or dynamic targets