Add base64 encoded user data passed from amazon deploy description
Reasoning:
We would like to pass User Data from an earlier stage in the pipeline to the deploy stage. This adds the ability to edit the pipeline json and add "base64UserData": "${userdata}" as a key value pair. The base64 user data is appended after the UserDataProviders sets up any default user data so it can be used to override these values.

It is base64 encoded because it is easier to handle as a serialized format when passing through json. Later on, a fancy UI could be built to add User Data as a text field. For now, advanced users can pass through the pipeline stages editing the json.