Docker Registry is now a Clouddriver provider.
The goal is to supply other providers with a way to retrieve cached image and
repository details. The next step is to actually cache the tags associated with
the repositories a user specifies. This setup works for any registry as long
as it conforms to the v2 API. Authentication is done using the token passing
scheme described [here](https://docs.docker.com/registry/spec/auth/token/).

The configuration looks like this:
```
dockerRegistry:
  enabled: true
  accounts:
    - name: my-docker-account
      username: lwander
      password: hunter2
      email: lwander@google.com
      address: https://index.docker.io
      repositories:
        - library/ubuntu
        - lwander/private-repository
```