Updates the FindImage stage to allow equivalent image lookup in regions with no ServerGroup.
Adds a `resolveMissingLocations` attribute to the FindImage stage that indicates if there is
not a server group for one of the required locations (zones or regions), to attempt to find
an image in those regions that matches the existing image.

Adds a global config property for the default behaviour of that flag if unspecified in the
request (defaults to false to preserve existing behaviour).

An image is only looked up if all the other find image results provided only a single
imageName - if the results returned multiple names the resolution gives up.

This enables the use case of promotion from test to production without requiring that
an image is deployed in test for each region that it ends up in production.

This requires that image naming ends up consistent (e.g. all regions need to be baked
at the same time).

This only currently will work for AWS as the clouddriver endpoint `/{cloudProvider}/images/{region}/{imageId}`
is only implemented on that provider.