Adding support for AWS AutoScaling Group Tags
Creates autoscaling groups with `tags` defined from a key-value map. Not sure
how y`all want to handle exceptions here, but if the user passes in more than
allowed the AWS API throws a LimitExceededException. Could handle it at a
validator side, but would rather not hard-code to a limit that might change.

Signed-off-by: Stephen Hephaestus Adams <heph@h3ph.com>