Handles duplicate bakes in cross region image lookup
On rare occasions the (Netflix internal) bakery will kick off two simultaneous
bakes in a region and end up producing two names for the results to resolve
uniqueness constraints, resulting in both
* `foo-120.2-h180.ffea4c7-x86_64-20160218235358-trusty-pv-ebs`
* `foo-120.2-h180.ffea4c7-x86_64-20160218235358-trusty-pv-ebs1`

In this case we accept either of those values and use the image find to search
against the base of `foo-120.2-h180.ffea4c7-x86_64-20160218235358-trusty-pv-ebs`. We
then iterate over any results that come back, and pull amis to match regions that were
requested but didn't exist in the search cluster.