fixes SPIN-1749: get package name when no jenkins task present.
The quip stage needs to get the package name to patch the instacne.  When using a Jenkins Job it is able pull this off the pipelines trigger object.

However Quip is often used with a Jenkins stage instead of a trigger resulting in an empty version resolving.

This fix also looks at both the buildInfo on the context and trigger to get the target deb version and then build the version name.