Azure server group caching bug fixes (#759)
When the user deletes the server group through the Azure Portal web site, it causes the Redis entries to be in a corrupted state. One side effect of it is Spinnaker continues to render "ghost" server groups that can be confusing.
Also fixed an issue with how deck is expecting load balancers and security groups associated with a server group to be passed (some core functionality is expecting a list rather than a string).