Stackdriver fixes.
If running on gce, you have to store metrics in the project of the managed
resources so force the project. But still need to configure the project if
you are not on gce (and thus dont have a gce managed resource).

When transforming [timer] metrics, lookup the hints for the original metric
name, not the transformed one.