chore(orca): Deprecate use of the global context (#1561)
This is the first step in getting rid of global context. Each stage can now define outputs that will be automatically resolved by the context of downstream stages. Ultimately global context is still used but it will trigger an Atlas counter on each read so we can start to track down remaining usages while not breaking anyone's pipelines.