feat(amazon): Support a generic proxy for static content stored in S3 (#1838)
In the short-term. this will allow a custom `deck` component to display
data that has been placed in an s3 bucket (in our case, by another team).

The proxy supports both static and adhoc configurations:
- `static`: A single bucket/object will be proxied (and temporarily cached)
-  `adhoc`: Any bucket/object will be proxied provided it matches a whitelist.
            The whitelist allows restriction on both bucket names and object keys.

Only buckets that Spinnaker has read access to are eligible for proxying.

The requesting user _must_ have access to the account containing the bucket/object
being proxied.