feat(artifacts): sets up artifact module (#2218)
My plan was to reuse the artifact work in appengine, but it would have
been too messy to refactor. It didn't expose any artifact-centric interface or
allow easy integration with other artifact sources without a fair bit of
extra work, so I rewrote in a separate module. T

he appengine GCS integration was left unchanged (with
the small change of moving the storage utils into a 'AritifactUtils'
class).