fix(jobs): Fix bottleneck around JobLocalExecutor.startJob (#2280)
The `startJob` method submits your request to a threadpool, which it
immediately starts polling. This isn't useful, and limits the number of
jobs that can be started to the number of threads supported by your
scheduler. As a result, submitting a lot of jobs at once causes
(needlessly) a lot of these jobs to timeout.

There is still a `sleep(500)` which doesn't seem to add anything in the
function, but didn't feel confident removing since other spots in
clouddriver make use of it.