chore(core): Consistency around `Millis` suffix on `task.*` values
We have a few situatations where task timeouts are overridable:

- tasks.waitForClusterTimeout (value is expected to be milliseconds)
- tasks.waitForUpsertedImageTagsTimeout (value is expected to be seconds)
- tasks.disableClusterMinTimeMillis (value is expected to be milliseconds)

The first two were added in the last 1-2 weeks, the last was merged in
late November.

My preference would be to align around the suffix for these sort of
properties, even if it might mean a breaking change right now.

Given we have slightly longer prior art, we could go with `Millis`
instead of the more traditional `Ms`.

If the original authors can reach consensus on `Ms`, we would be just
as happy to go with that.