fix(core): prevent stages evaluating target server groups twice
The problem previously was that we'd end up calling the aroundStages when planning before stages and then again when planning after stages (following recent changes to do that just in time). That meant the target resolution stuff happened twice and the second time it might not be valid (because of what the stage did to its target). I've changed it so that everything happens in before stages. Before it was transforming its own context to deal with a single target then spawning copies to deal with each other target. It also had to do some complex shenanigoats to weave in before and after stages of each target stage. Now the top level stage is a no op and just creates child before stages to deal with each target. Those child stages can worry about their own before and after stages so the whole thing is just a lot simpler.