refactor(persistence): Unify Jedis/Dynomite Repos (#2070)
Having separate Jedis vs. Dynomite execution repositories posed a number of
problems around live migrations in a dual repository configuration.

This was partially due to the Dynomite repository using a different
scheme for key names, in order to utilize Dynomite hash sets, I.e.
[pipeline:abc-123, pipeline:abc-123:stageIndex] turns into
[{pipeline:abc-123}, {pipeline:abc-123}:stageIndex]. To support
live migrations, we need to be able to find executions in either
the current or previous store regardless of type or key names.
This implements that logic, as well as the ability to generate key
names based on the class type of each configured RedisClientDelegate.
This should enable Redis <---> Dynomite migrations in either direction.