perf(aws): Optimize the retrieval of all security groups (#2539)
When `includeRules` is `false`, the cache key for a security group
actually includes all relevant information.

We _do not_ need to retrieve all objects and can instead just parse
keys and manually construct an `AmazonSecurityGroup`.

There is _some_ chance that the set of identifiers includes some
security groups that do not actually exist. I believe this to be of
minimal risk and the performance benefits outweigh the potential for
issue.

The return types in `SecurityGroupProvider` were changed from `Set` to
`List`. It turns out that converting a list of 20k items to a set takes
a non-zero amount of time.