feat(qos): Adding initial naive Quality of Service module (#2164)
We've been experiencing production pain in the form of higher bursty execution
load than what Orca's downstream services are able to take if offered in a
firehose fashion. This PR offers a framework (albeit simplistic first impl)
for enforcing QoS of executions.

The basic premise is that a new extension hook is added when executions are
initially saved into the system, they'll potentially be set with a BUFFERED
status, which will prevent StartExecutionHandler from seeing it, and thus
beginning any work.

A BufferStateSupplier is implemented (currently based off system-wide active
executions) that can flip the system from INACTIVE to ACTIVE. In the case of
this basic implementation, if a threshold of active execution is surpassed the
QoS system is set to ACTIVE.

Once ACTIVE, an ExecutionBufferActuator listening to the new
BeforeInitialExecutionPersist event will start passing the execution through
a pluggable & configurable set of ordered BufferPolicies. A BufferPolicy is
responsible for determining an action on the execution, either to ENQUEUE
BUFFER. If ENQUEUE, the execution will be dropped onto the queue immediately,
BUFFER will set the execution's state to BUFFERED.

On the other end, an ExecutionPromoter is always running on an interval,
regardless of BufferState, to promote BUFFERED executions into NOT_STARTED
(and thus enqueued). The ExecutionPromoter, similar to the
ExecutionBufferActuator, will pass all candidate executions through a list
of ordered PromotionPolicies. These Policies will apply heuristics for
filtering and ordering what executions will be promoted.

In both forms of Policy, a Policy can force an action if it decides that it
knows best. If multiple Policies perform a force, the highest precedence
Policy will win.

---

Currently, the implementation is very naive and basically on/off. Future work
will be done to implement application & execution criticality to inform
policies. I'd also like to profile individual clouddriver clients as another
form of BufferStateSupplier, so we can back off on work we're sending to
clouddriver if orca starts getting higher latencies or increased error rates.

There will be additional further work in attaching system notifications to
executions so that Orca can propagate QoS actions up to end users via the API
and UI so people aren't left wondering why their execution isn't immediately
starting.

This concept of execution buffering was chosen over something implemented
on the queue so that we can ensure in-flight executions are finished as
quickly as possible, rather than slowing all executions down. It's possible
this QoS system may grow more sophisticated in the future to tune QoS of
in-flight executions based on criticality, but that's a lot more work and a
little more dicey.