fix(fiat): Transitional support for fiat migration.
Continue to propagate X-SPINNAKER-ACCOUNTS header when
fiat is enabled, allowing a live upgrade into fiat.

Removes auth filtering on /credentials endpoint.
Credentials include an authorized flag which can be used
for filtering choices in the UI, but account metadata is
required for proper UI rendering even for accounts that
the user does not currently have permission for.