fix(provider/amazon): improve instancetype caching
Switches AWS Instance Type caching to process the amazon billing api
offerings file instead of making an API call to reserved instance
offerings.

Polls this endpoint on the regular caching interval, but only
processes it if it serves up a different etag value than the
previous cycle.

The billing endpoint is listed as us-east-1 but is actually fronted
by CloudFront so in theory shouldn't be an issue to hit from any
specific Spinnaker installation.