fix(cron): fallback for cron triggers with null id (#270)
* fix(cron): fallback for cron triggers with null id

If pipeline configurations with null cron trigger ids somehow went past
front50, we would have "flagging" registrations: new triggers would be found,
then on the next pass of the PipelineConfigsPollingAgent we would find
scheduled actions with no corresponding triggers (since they don't have)
a matching id. We would then deregister then, only to register them again
right away, and so on.

With this change, we no longer need a persisted id for cron triggers in
the pipeline configurations (which is bound to be messed with, deleted,
copied, mangled...). This essentially deprecates the stored trigger ids in 
pipeline configs in favor of the dynamically generated ones

generateFallbackId() hashes the toString() representation of Trigger,
which should work fine as long as it includes at least the parent
pipeline id and the cron expression