fix(core): Remove the circular dependency on `clouddriver` (#234)
This PR makes use of the locking support in `kork` and removes a
dependency on `clouddriver:cats-redis`.

Redis usage has migrated to the `RedisClientDelegate` which in theory
opens us up to using dynomite at Netflix.

BREAKING CHANGE

Introduces a new config fiat.writeMode.syncFailureDelayMs as a means
of controlling the lock delay on failure vs success of a sync. This was
previously managed by redis.poll.errorIntervalSeconds.