fix(entitytags): Don't set _type field on Elasticsearch documents (#2913)
Storing multiple mapping types in the same index is deprecated in
Elasticsearch 6 and will be removed in Elasticsearch 7. Add a config
parameter indicating that only one mapping type per index is supported
(defaulting to false, for backwards compatibility). If that parameter
is set, all documents will have their type set to the config parameter
mappingTypeName, which defaults to Elasticsearch's recommended
value of _doc.