refactor(core): Create new endpoint to manually trigger with Echo (#600)
* refactor(core): Create new endpoint to manually trigger with Echo

This change adds a new endpoint to trigger a pipeline that will send
the request through Echo rather than directly to Orca.  By sending
the request through Echo, processing steps that happen for automated
triggers will also be applied to manual triggers. Eventually we'll
migrate existing workflows to go through Echo, but for a safe rollout
we're leaving the existing endpoint unchanged.