Add echo manual trigger handler (#341)
* refactor(core): Create handler for manual triggers

This creates a simple handler to process manual trigger events
and start the corresponding pipeline via a call to Orca. This is
the first step in having Echo do the same processing (ex: inflating
artifacts) on manually-triggered events as on automatically-triggered
events. This code will not actually run until a corresponding change
to gate actually sends manual events to echo.

* refactor(core): Push metric recording to base class TriggerMonitor