Fix duplication instance health (NLB) [GCP] (#3023)
* fix(provider/google): Fix empty server group with Ssl/Tcp load balancer

* fix(provider/google): Do not add duplicate instance healths (NLB)

Changes:
- Skip health status for a different address
- Only add one health resolution per load balancer, by using a set
- Revert the de-duplication workaround in bc4ed8f0

A NLB setup involves a target pool, along with:
- X addresses
- Y forwarding rules ("load balancers") per address
- Z instances

There are X * Y load balancers. Before this change, within each load
balancer, there would be X * Z * Z instance healths, when it should
be just Z.

With 8 addresses, 3 forwarding rules per address, and 60 instances,
there would be 8 * 3 * 8 * 60 * 60 = 691200 instance healths. G1GC, on
a powerful machine with a 32GB heap, will not be able to keep up with
such load, and will repeatedly fallback to serial full GC, causing
long STW pause that lasts for many seconds.

The de-duplication masked the problem, but doesn't help the GC load.

Anyway, the API design for NLB is a little unintuitive, such that:
1. The health check is tied to the target pool.
2. We will make Z API calls to "targetPools.getHealth", once for each
   instance.
3. Each response will contain X health statuses for the instance, one
   for each address.
4. We will then populate the `tpNameToInstanceHealthsMap` from the
   responses.
5. To populate the instance healths for each load balancer, we will
   dip into the map X * Y times.

The duplication could happen, because:
- we didn't deal with the X health statuses correctly in step 3.
- we dipped into the map X * Y * Z times in step 5.

* fix(provider/google): All LBs of the same class had the same hashcode

This fixes a subtle bug where both the parent and the child have the
`@Canonical` annotation, and the distinctive attributes are at the
parent (e.g. name and ipAddress).

This is a prerequisite fix, before we can tackle the problem of
duplicate entries in `GoogleLoadBalancer.healths`.

* fix(provider/google): Do not add duplicate instance healths (others)

Similar to the fix for NLB, but for Http/Ssl/Tcp/Internal load
balancers. Also, unlike the address count * instance count multiplier
for NLB, the impact here is much lower, as the multiplier is just the
backend count.

* fix(provider/google): Do not pass duplicates to `setTargetPools`

This is just to make the output of `instanceGroupManagers.get` to look
nice, the duplicates do not seem to cause any problem.