feat(core): Add pipeline post-processor functionality (#380)
* refactor(core): Convert echo-core to Java

Echo-core is almost completely java, with the exception of
two easy-to-convert files. Finish the conversion and change
the top-level source directory from groovy to java.

Also, add some explicit gradle dependencies.

* feat(core): Add support for requesting build info

Add the required infrastructure for echo to request
build info from Igor.

* feat(core): Add pipeline post-processor functionality

Add functionality to post-process pipelines before sending
them to orca for execution. In particular, add a post-processor
to add build info to the trigger, and a (not-yet-implemented)
post-processor to inflate artifacts using a Jina template.

* feat(core): Add igor endpoint to halyard base config

* style(core): Address code review comments

* refactor(core): Don't rely on autowiring for ordering

Add functionality for ordering pipeline post-processors that
doesn't depend on Spring autowiring the post-processors in the
order specified by @Order.

* style(core): Address PR comments

* Clean up EchoRetrofitConfig
* Add retry to igor communication