feat(core): Apply artifact filtering in Igor (#321)
* feat(core): Apply artifact filtering in Igor

BREAKING CHANGE: This is a breaking change for users who
have manually set BuildArtifactFilter.maxArtifacts or
BuildArtifactFilter.preferredArtifacts as environment variables
to Orca. As the filtering logic is being moved to Igor, these
environment variables need to be set on the Igor process. This
change will also cause the results of calling the Gate endpoint
"/v2/builds/..." to get build status to return a filtered list
of artifacts.

All current calls to get the build info for a particular build
post-process by filtering the artifacts with BuildArtifactFilter.
Rather than require all callers to post-process, move this logic
from Orca to Igor so that the API returns already-filtered
artifacts.

* refactor(core): Convert artifact filtering to Java