feat(webhooks): Allow a custom trust store for webhooks (#2517)
* refactor(webhooks): Convert webhook config to Java

Before making changes to these two classes, convert the
existing classes to Java.

* refactor(webhooks): Use explicitly configured okhttp client

Change outgoing webhooks to use an expclitily-configured okhttp3
client rather than use default settings. This will allow future
configuration of the trust store.

* feat(webhooks): Allow a custom trust store for webhooks

This change allows users to configure a custom trust store
that will be used for webhooks; it will be used in addition to
the default trust store.  (The use case is where a user has
an additional CA they would like to trust in addition to the
standard CAs.)