fix(traffic-guards): handle server groups in a namespace
If a TargetServerGroup has a 'namespace' property, then its getLocation()
returns a location tied to that namespace (regardless of region, zone, etc.)

If the location passed in verifyTrafficRemoval is different (e.g. tied to
a region) then we want to be explicit in the location type to use for the
lookup/filtering.