fix(traffic-guards): be more permissive with no-op removals
Instead of failing on null/empty server groups to remove, consider that
a no-op and exit early.

For instance, it can happen if you request a shrinkCluster operation
on multiple locations but there are no server groups in one of the locations.

It seems safe and simpler to relax that constraint rather than assume that
every caller will perform the required validation.